# services/core/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CreateAIPlanRequest:
    properties:
      description:
        type: string
      employee_id:
        type: string
      topic:
        type: string
    required:
    - employee_id
    - topic
    type: object
  dto.CreateAIPlanResponse:
    properties:
      id:
        type: string
    type: object
  dto.CreatePlanRequest:
    properties:
      description:
        type: string
      employee_id:
        type: string
      title:
        type: string
    type: object
  dto.CreatePlanResponse:
    properties:
      id:
        type: string
    type: object
...
  dto.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  dto.ListPlansResponse:
    properties:
      plans:
        items:
          $ref: '#/definitions/dto.PlanResponse'
        type: array
    type: object
...
  dto.PlanResponse:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      creation_type:
        type: string
      description:
        type: string
      employee_id:
        type: string
      generation_status:
        type: string
      id:
        type: string
      progress:
        type: integer
      status:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
  dto.PlanWithTasksResponse:
    properties:
      plan:
        $ref: '#/definitions/dto.PlanResponse'
      tasks:
        items:
          $ref: '#/definitions/dto.TaskResponse'
        type: array
    type: object
...
  dto.TaskResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      plan_id:
        type: string
      position:
        type: integer
      status:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
...
  dto.UpdatePlanRequest:
    properties:
      description:
        type: string
      title:
        type: string
    required:
    - title
    type: object
...
host: localhost:8080
info:
  contact:
    name: Mandzhonson
  description: REST API для SkillTracker.
  title: SkillTracker API
  version: "1.0"
paths:
...
  /employee/plans:
    get:
      consumes:
      - application/json
      description: Возвращает список планов текущего сотрудника
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.PlanWithTasksResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить список планов сотрудника
      tags:
      - Employee
  /employee/plans/{plan_id}:
    get:
      consumes:
      - application/json
      description: Возвращает детальную информацию о плане сотрудника
      parameters:
      - description: ID плана
        in: path
        name: plan_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PlanWithTasksResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить план сотрудника по ID
      tags:
      - Employee
...
  /manager/plans:
    get:
      consumes:
      - application/json
      description: Возвращает список всех планов, созданных менеджером
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ListPlansResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить список планов менеджера
      tags:
      - Plans
    post:
      consumes:
      - application/json
      description: Создает новый план развития для сотрудника
      parameters:
      - description: Данные для создания плана
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreatePlanRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CreatePlanResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Создать план
      tags:
      - Plans
  /manager/plans/{plan_id}:
    delete:
      consumes:
      - application/json
      description: Удаляет план и все связанные задачи
      parameters:
      - description: ID плана
        in: path
        name: plan_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Удалить план
      tags:
      - Plans
    get:
      consumes:
      - application/json
      description: Возвращает детальную информацию о плане с задачами
      parameters:
      - description: ID плана
        in: path
        name: plan_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PlanWithTasksResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить план по ID
      tags:
      - Plans
    patch:
      consumes:
      - application/json
      description: Обновляет название и описание плана
      parameters:
      - description: ID плана
        in: path
        name: plan_id
        required: true
        type: string
      - description: Данные для обновления
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdatePlanRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Обновить план
      tags:
      - Plans
  /manager/plans/{plan_id}/archive:
    patch:
      consumes:
      - application/json
      description: Переводит план в статус archived
      parameters:
      - description: ID плана
        in: path
        name: plan_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Архивировать план
      tags:
      - Plans
...
  /manager/plans/ai:
    post:
      consumes:
      - application/json
      description: Генерирует и создает план развития с помощью ИИ
      parameters:
      - description: Данные для генерации плана
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateAIPlanRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CreateAIPlanResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Создать план с помощью ИИ
      tags:
      - Plans
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"